Add support for Open ID Connect (OIDC) identity provider
Open ID Connect is an authentication protocol build on top
of OAuth2. It is used for example in Azure AD and Keycloack

django-allauth unfortunately does not support the generic OIDC.

However, Mozzila developed a OIDC library for django called
`mozilla-django-oidc`

The integration is entirely optional and nothing is imported
from `mozilla-django-oidc` unless it is activated with `AUTH_OIDC`
key in YAML configuration files.