lava-dispatcher-host: return empty list for empty mapping
Loading an empty string as YAML returns None, and in this context need
the result to be a list because calling code assumes it to be one.

Now the question is how exactly we are ending up with an empty
usbmap.yaml file. This might be happening in situations of system
streess, or maybe due to a race condition. In any case, this ensures
that the return value at least makes sense.