lava_server: make it possible to enable all login options at once
If a social account login was enabled, that would force using the login
view from from django-allauth, instead of the existing Django login
view, with the template overriden in lava_server. Instead of giving
priority to the routes, we instead list the social auth providers in the
existing LAVA login page.

Add logos for gitlab, github, and google because those are the most
likely to be used by people.