DockerDriver: ensure unique container names
This avoids a race condition where two commands are run in quick
succession, and the second fails because it cannot start a container
with the same name as the one before. Containers are cleaned up right
after the command finishes, but there is a possible race condition
between the removal of the first container (which is done in the
background by the docker daemon), and starting a new one.

We mitigate this be ensuring that each command run through docker has a
unique name, by appending a UUID to it.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>