connections: serial: allow devices with no `commands` config
This allows for example to use minimal boot actions with QEMU devices.

Since the check for the existance of commands was being done after it
was already used anyway, this can't introduce a regression. If commands
were not defined before, this would cause a crash at the beginning of
the job.

Now, if there are no defined commands, validation will skip all further
checks, and the ConnectDevice action will only work if there is already
a serial connection made by a previous action.