Check image's architecture without running it
No need to check the output of the "arch" on the container command line,
we may have been given an image that never reaches a shell. If this
docker container runs indefintely then LAVA cannot stop it. This can
lead to LAVA being unable to parse stderr as it is still being written
to once the job has ended which can in turn lead to worker crashes.

Instead just read metadata given from "docker inspect".

Change-Id: I1c00eb7647a5dbcc916862491c9781229b34c834
Signed-off-by: Dean Birch <dean.birch@arm.com>