docker test shell: don't override common connection
In order for the shared lava test shell code to use the docker container
as connection, we used to *override* the connection in the "common"
namespace with the docker one. This causes feedback from the DUT to not
be read while the tests are running, what makes debugging harder. For
example, if the tests timeout, the job fails right away and the feedback
from the DUT never gets to the logs.

Instead, create a temporary namespace called `docker-test-shell`, save
the connection to it, and tell lava test shell to use that namespace
instead of the default `common`. Before exiting, delete that namespace,
so this doesn't affect subsequent actions.