boot: qemu: allow host bridge network in docker
To allow qemu devices both inside and outside to operate mosty
identically, especially with respect to netdevice configuration, the
docker network environment should match the host network environment.

To achieve that tell docker to set the network to host and add the
NET_ADMIN capability. On top of that to support the qemu bridge
netdevices add the tun device from the host.

Fixes: #548

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>