Give error_msg and error_type
It is currently impossible to get error_msg/error_type for a specific
job.
The only way to get it, is via scheduler.jobs.list() but this function
need to find the job via relative offset from last job, and so it is
hard to get it.

This is an exemple on how to give error_type/error_msg for a specific
job.

Closes: 555
Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>