auth: Fix two authorization issues for device (type) visibility
First one is on worker detail page while looking at transitions, a
non-authorized user can see device transitions for devices he's not
supposed to.
Second one is the device type health history; users are able to view
the whole page they're not supposed to. Also on the same page,
transitions are shown to non-authorized users.

Signed-off-by: Stevan Radaković <stevan.radakovic@linaro.org>