Substitute parameters within install:git-repos
At present, a parameter can only be used within install:git-repos if
it is the entire value of the target, so we need a dedicated parameter
for the specific repository URL in a test, for example. This becomes
very inconvenient in larger test sets, for example if we need to
create a parameter for every repository, because the repository URL
needs to contain short-lived tokens that must be substituted for.

This goes to some pains to only take action on strings on which there
is a valid parameter substitution present. This is with the goal of
minimising possible regressions from introducing this syntax.

Signed-off-by: Ed Smith <ed.smith@collabora.com>