Run django migrations as the correct target user
Running them as root leaves blank unwritable logs if they don't
already exist, as is the case for the docker image. It seemed safer to
run the commands as the correct user (to avoid future permission
problems) than to simply bughunt the files that are currently created
that have the wrong owners.