Add an API for setting the groups a user is a member of
Lava uses XMLRPC for its system management endpoints, which is secured
via HTTP basic authentication. This adds `system.set_user_groups`
taking a string email address for the user, and an array of string
group names which will be the exact set of groups of which the user is
a member.

This endpoint requires a token from a user with staff permissions in
order to access it.