Add ava and base-edk2 device types
base-edk2 is a very bare bones device type, that only supports the
minimal boot option, and no deploy option. Actual deploy for now needs
to be done via the UEFI Shell, either by a UEFI application that does
that, or by booting a Linux image that writes to any disk on board.