Fix a huge performance issue when parsing kernel boot log
Prior to this patch, booting a really verbose kernel and filesystem under qemu
was really CPU-intensive for lava-run. In some specific circumstances, lava-run
was too slow to consume qemu output, making qemu to hang.

This is due to the way lava uses pexpect. pexpect was configured to scan the
full buffer (since the start of the boot) for each kernel error messages.

Now pexpect is configured to only look at the last 4000 characters in the
buffer, making the parsing quicker and time-bounded. In order to still capture
long kernel messages, the mesages has been cut in two (start, end). Where the
start is always shorter than 400 characters. When matching the first part,
pexpect is reconfigured (temporarily) to look at the full buffer for the
remaining part.

With this patch, booting a qemu-arm64 verbose kernel/rootfs is now similar with
or without lava.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>