Add schema support for docker login
Most actions rely on the same utility function to describe the docker
parameters they take in their schema. This extends that function so
that, optionally, an extra stanza for login details is recognised. By
default, the stanza is not recognised, it must be opted into.

The reason this stanza is optional is that it will take time to
implement docker login support in all the necessary places. This
allows us to implement it for a subset of LAVA actions, whilst keeping
the schema checking accurate for actions which do not yet have login
support.