utils: compression: untar_file: set dereference
Overlaying modules into a debian rootfs where '/lib' is a symlink to
'/usr/lib' will end up in a 'Kernel panic' since init can't be found.
What happens is that LAVA untar the modules in '/lib/modules' which
makes the '/lib' a directory.
To prevent this to happen is to tell tar to follow the symlink and untar
the modules there, this can be done with 'dereference=True'.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>