lava_common: avoid data sharing between worker and docker worker
As worker and docker worker use the same data directory by default for
db and log files, when both workers are installed and running on the
same OS for some reasons, job started by docker worker will be updated
as finished by the other worker in its next ping cycle. That is
because the lava-run process ID for the job exists in docker container
but doesn't exist on the main OS.

When the above issue happen, jobs finished correctly are marked as
incomplete. That is because when the jobs reported as finished to
lava server, the lava-run process is still running on the docker
container and result.yaml file is not generated yet and job result
defaults to fail.

Using different data directory by default can avoid the potential
confusion.

Signed-off-by: Chase Qi <chase.qi@linaro.org>