Add a REST API alongside the XMLRPC API for group management
There are two largely parallel APIs for interacting with LAVA,
XMLRPC and REST. This adds the new methods

  set_user_groups
  set_user_active
  get_all_users

to the REST API from the XMLRPC API.

These are all email based, to match XMLRPC, although it's possible
we'd like to transition over to using usernames instead in the near
future. This brings both APIs to the same point; any future change
can be applied to both together with little difficulty.

Signed-off-by: Ed Smith <ed.smith@collabora.com>