lava_server: Add a placeholder to table search pointing to help
Previously the table attempted to put a value in to the search box
that would always fail because of programming error.

This resulted in empty search bar. The new placeholder value is
in the background and tells user to press the question mark to
get help on available methods and fields.