tests/lava_server: fix tests for Python 3.11
Under python 3.11, pkgutil.resolve_name(), called by
unittest.mock.patch, will raise an ValueError if the name to be resolved
(the `target` argument to patch()) contains dashes (-). So we need to 1)
refer to the function to be patched indirectly via the current module,
because Django command modules do contain dashes; and 2) rename the test
file itself so it no longer contains a dash in its name.

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>