Improve handling of docker login interactions with remote hosts
Each job gets a pristine HOME directory on the worker, which means
that when we call `docker login`, the credentials are not available to
any other job. That in turn means no other job without valid
credentials can pull that image. We prevent no-pull usages of docker
when we permit private images to be pulled.

The situation becomes more complicated when we talk to a remote host.
Calling docker login may create stored credentials on the remote. In
any case, we're in general unable to prevent other systems from also
talking to that remote, and making use of the private images we have
pulled to it.  Nor are we capable of wiping those images after we have
used them (this would not fix the issue, but it would make the
vulnerability window shorter).

For these reasons, we do not run docker login on remote hosts. There
are three unique docker implementations in this codebase, and the
above is enforced by:
- `deploy:docker` throw an error when requesting `remote` and also
  `login`. This is a job error, because these things can never work
  together.
- `boot:fvp` does not currently support using a remote host, or
  any other extended docker options that might be used to insert
  `--host` or similar into the command line.
- `utils:docker` is the workhorse for most actions. It does support
  arbitrary pre-command docker options, which means the user can
  specify `--host` and similar. Securely parsing all possible forms
  of these arguments to determine whether this has happened is deferred.
  Instead, we omit those arguments when calling login (issuing a warning).
  That means we will not login on a remote. The job may then fail,
  but no security breach will occur.

Signed-off-by: Ed Smith <ed.smith@collabora.com>