fix django foreign key model access
In Django >= 2.0 the way to access a foreign key model has changed. The
current `lava_results_app` model is still returned with the `rel.to`
attribute. This generates `AttributeError`s in LAVA when running queries
which need to access such keys (e.g. `requested_device_type` or
`actual_device`). Fix it by accessing the field directly.