Introduce annotate_int_field_verbose function
It will be used to convert the int field values to the verbose
names.

Example:

Device.health -> 4 -> "Maintenance"

Usuallt the get_FOO_display can be used to get the verbose name
but that requires fetching the entire model which is not available
when values() or values_list() is used.

Instead the When-Case-Values construct can be used to annotate
with verbose name the values on the database. The
`annotate_int_field_verbose` will create those annotations using
the Django field metadata.

This technique was already used in 2be1e2b207f93f0a82306fabe9f8b302f377a6ab
but is now in a separate function and can be applied to other
tables.