lava-docker-worker: service: set the timeout to 90s instead of infinity
The current setting (infinity) prevent to reboot the system as we're
waiting indefinitely to stop the service.

The default TimeoutStartSec is already 90s, so set also TimeoutStopSec to 90s
as it's a reasonable time to wait for the service to stop.

Use TimeoutSec to set the value.

Note: TimeoutSec is a shorthand for configuring both TimeoutStartSec= and
      TimeoutStopSec= to the specified value.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>