Relicense LAVA to GPLv2 or later
Before this patch, LAVA code was licensed under GPLv2 (or later) or AGPLv3,
depending on the python module. This commit relicense all the code to GPLv2 or
later.

Approval from the relevant contributors has been requested and granted before
hand. Contributors will confirm the aggreement by replying to the gitlab merge
request. The commit will be updated to add the relevant Acked-by flags.

Acked-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>
Acked-by: Jan-Simon Möller <dl9pf@gmx.de>
Acked-by: Karsten Tausche <karsten@fairphone.com>
Acked-by: Kevin Hilman <khilman@baylibre.com>
Acked-by: Konrad Scherer <Konrad.Scherer@windriver.com>
Acked-by: Philippe Mazet <philippe.mazet@nxp.com>
Acked-by: Sjoerd Simons <sjoerd@collabora.com>
Acked-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Acked-by: Tim Jaacks <tim.jaacks@seco.com> (formerly known as Garz & Fricke GmbH)

Signed-off-by: Rémi Duraffort <remi.duraffort@linaro.org>