test_schroot_params: Replace the use of deprecated `run_command`
Instead use new `parsed_command`.

The `parsed_command` saves the output to `.results["output"]`.
This means the action will no longer become invalid when a
subprocess raises an non-zero exit code. Instead of checking
`.valid` check the return code is equals to zero.

Also remove check of schroot name. If the name of the schroot
changes in `ssh-deploy.yaml` this will cause errors. There is
no point in checking for a specifc specific name.