Disable worker auto-registration by default
This option has security implication if not configured correctly.

The default value WORKER_AUTO_REGISTER_NETMASK allows localhost
to register workers, however, if you have a HTTP proxy server
LAVA will see the proxy server IP address which might be a localhost.
In this case anyone will be able to register a worker and obtain
the token.