Remove use of deprecated `zmq.utils.strtypes` from lava-scheduler
That module was useful in Python 2 to 3 transition era but
nowadays the string type is known in advance.

`settings.EVENT_TOPIC` is always going to be `str` but
`setsockopt` accepts only `bytes`.

`recv_multipart` always returns list of `bytes` therefore
`topic` should be decoded but `json.loads` can accept `bytes`.
This also means we can avoid decoding unnecessary message fields
and avoid loading JSON if the topic is not matching.