Remove `simplejson`
The module is equivalent to standard library module `json` but is
actually slower.

See: https://github.com/kytta/python-json-benchmark

Replace it with a mix of `json` and `JsonResponse` (from django).

`JsonResponse` requires `safe=False` when an array is returned
as response because before ECMAScript 5 (released in 2009) the
array constructor was vulnerable to poisoning. However, all modern
browsers support ECMAScript 5 and higher.