tests: Test notification callback JSON generation and requests
Before the notification tests would only test the values but not
the JSON generation behavior or requests library call.

Now the `invoke_callback` function will be called while `requests`
module is patched. The mock is later checked if it was actually
called.

The JSON is generated in to temporary directory. Only its presence
is checked in the directory.