schema: allow any object type for job metadata values
Widen the range of allowed types for job metadata values in the
original schema to align it with the new schema {str: object}.  This
change is backwards-compatible and allows objects such as lists or
dictionaries to be passed in the job metadata rather than only strings
and integers.  For example, this is now possible:

  metadata:
    api_config:
      name: staging.kernelci.org
      url: https://staging.kernelci.org:9000
      version: latest
    node_id: 645372a98fec6b4207361e03

Update unit tests accordingly with metadata keys as int or str and
values as various object types including a list and a dictionary.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>
Fixes: https://git.lavasoftware.org/lava/lava/-/issues/610