docker-worker: don't use cache when building custom image
Use docker build `--no-cache` by default to force a clean build when
restarting docker worker service. This is to make sure the latest
dependences defined in Dockerfile are built into the custom image,
typically when the depends are pull from git repo or downloaded from
internet.

Add `--use-cache` to allow to override the default behavior.

Signed-off-by: Chase Qi <chase.qi@linaro.org>