Fix filtering by permissions returning duplicate rows
The issue was that Django would put LEFT OUTER JOIN in to the
resulting query. This caused duplicate rows when filtered object
could be accessed by a given user through multiple groups.

Now filtering by users permissions is separated in to a subquery.
This also gives a small performance boost.