Avoid loading device type columns for job tables
The device type is only used to create a link to the device
type details page. Only a primary key is required for this and
it is already available under `requested_device_type_id` attribute
of the job.

Avoiding loading the device type columns will reduce amount of
buffers used or disk reads.