Avoid loading device columns for job tables
This is equivalent to commit 1ce10d330646a6e6e83b93782fd8ad23dedffe6e
but for actual device columns.

The device type is only used to create a link to the device details
page. Only a primary key is required for this and it is already
available under `actual_device_id` attribute of the job.

Avoiding loading the device columns will reduce amount of buffers
used or disk reads.