Use annotate for job failure metadata instead of micro queries
Instead of running a query for each failed job metadata use
annotate to include the metadata in a single query.

This significantly improves performance.

Apparently `failure_comment` is a pretty rare field so metadata
is almost always used so there is no point in loading it conditinally
instead of always loading it.