Fix actual device column render function not being called
The issue is that the column was renamed from `actual_device`
to `actual_device_id` in 30345d14f57332b5dee083af4f21ab9728b8db0c
to avoid database lookup. However, the rendering function was
supposed to be renamed as well as it has to follow the
naming of column.