Remove `terms_data` from LAVA tables
The only purpose it could have served is displaying help.
However, none of the tables set `Meta.terms` dictionary.
The case sensitive search help is already covered by directly
using the `Meta.searches` attribute.

Therefore the entire `self.terms` dictionary can be removed and
the general search can be replaced with a single variable.
This variable does not need HTML escape because it would not
be used in rendering at all.