device-types: rk3588-rock-5b: remove arbitrary test_character_delay setting
In 4f5ab9f7b773 the device-type for Radxa Rock 5B was modified to add
additional test_character_delay setting to bring the template in line
with rk3399-rock-pi-4b.jinja2. It turns out (from additional testing) that
this setting is not required on this device. Remove the additional
test_character_delay setting.

Fixes: 4f5ab9f7b773 ("device-types: rk3588-rock-5b: Update for mainline U-Boot")
Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>