Increase fetch depth for Gitlab CI
Otherwise the `version()` function will fail when running CI
as it relies on latest git tag being available.

Increase it to 100 as it should be enough to fetch until last
tag.

Previously the container image would contain a LAVA git mirror
that would be reused when running CI. The tags could be used
from that repository. Now CI fetches an empty repository.