lava_dispatcher: remove hard-coded pexpect searchwindowsize
The current logic that setting searchwindowsize to twice the size of
maxread should be pretty safe as pexpect read and search in serial.

The ShellCommand level maxread size defaults to 2000. connect-device
action overwrites the default with the spawn_maxread defined in
`base.jinja2` that defaults to 4092. The both defaults look ok as the
strings we search usually are small.

I think we just need to remove the hard-coded searchwindowsize to make
sure that it always is twice the size of maxread.

Signed-off-by: Chase Qi <chase.qi@linaro.org>