Pass `is_favorite` context variable to `job_definition` template
Depending on if the job is already favorite the button text
should change from "Add to favorites" to "Remove from favorites".

This behavior is already present but lacks the necessary context
variable to be properly working. Reuse code from `job_details`
view by separating it in to a function and using this new function
across all 3 views.