Enable REST API browsable filters rendering
It was disabled because of performance issues, however, same issues
can be fixed by changing the filter widget.

There is only only filter that might have rendering issues which
"last_health_report_job" of the Device filters. It tries to
render a choice box with all the health jobs. Replacing its widget
with the NumberInput fixes the performance issues.