Remove Debian 10 RelatedFilter argument name workaround
In the djangorestframework-filters version 1.0.0.dev0 that is
shipped starting from Debian 11 the argument `name` has been
renamed to `field_name`. This is an incompatible change so
to support both Debian 10 and 11 a workaround was added to the
`lava_server/compat.py`.

However, the minimum support Debian version is now Debian 11 so
this workaround can be removed and instead the `field_name` can be
used directly everywhere.

Also remove the `RestFrameworkFilterBackend` workaround and
subclass it directly to disable the filters' HTML form rendering.