Rework ResultsTable to not display restricted visibility jobs
Currently ResultsTable would display rows for all TestSuites
including the ones from the jobs you are not supposed to see.
In those rows it would simply use the `render_` functions to not
render certain values. However, that still leaked the job numbers
and their test suites names and showed exactly which jobs you
were not supposed to see.

Instead rework the table to start with the TestJobs and filtering
them by visibility. After that use values to make a join on the
TestSuites and use annotated subqueries to count the test cases.
This has better performance and no longer shows the restricted jobs.

However, the TestSuite materialized queries cannot be adapted for
this without breaking the existing materialized views because
they use the TestSuite for base query. Instead copy paste the
existing ResultsTable code for it and decouple it from new
ResultsTable. Drop the render visibility code as Query adds
visibility filtering already.