Refactor `get_restricted_job`
First, it was placed in the `lava_scheduler_app.views` module
even though it was not an actual view. Move it to a `@classmethod`
of the `TestJob` model just as `get_by_job_number` is.

Second, it awkwardly imported and used a function `check_request_auth`
from the `lava_results_app.utils` which created a circular dependency
in `lava_results_app` therefore that app did not make use of
`get_restricted_job` and instead reimplemented it everywhere
it wanted to check authentication.

One potentially significant difference is that `get_restricted_job`
will no longer check for `AuthToken`. This is because both REST API
and XMLRPC have their own mechanisms to replace request's user
based on the token. (in `lava_rest_app.authentication` and
`linaro_django_xmlrpc.views` respectively)