lava_dispatcher: disconnect device after uuu boot action
If a device serial is not disconnected after a `UUUBootAction`, the
generated U-Boot output will remain stored in the pexpect buffer
process. This output can then cause problems on subsequent actions. For
instance, in a following `UBootAction`, an unexpected prompt may be found before the
actual U-Boot prompt appears. Since the serial connection is not used by
pexpect during `UUUBootAction`, close it at the end.