Add YamlField for annotating querysets
It is useful if one of the querysets text fields contains a YAML
string would need to be parsed with `yaml_safe_load` to access
a specific YAML field. Instead of manually checking if the field
has been parsed just let Django do the parsing when executing
query.

To use add `output_field=YamlField()` the annotated field.