Clean-up Pipeline's add_action timeout value selection
First, change the algorithm that selects the last available value
of the passed list of dictionaries. Instead of always iterating
over all dictionaries just start from the end and return the
first found value. This will return much faster if a value was set.

Also there is no need to use get the latest dictionary value for
the action block timeouts as there will only be a single dictionary.