Make Notification's `test_job` column its primary key
Because each Notification has a one to one relation with a TestJob
there is no point in having a separate primary key `id` column.

This improves IO performance and storage space as less data is
needed both per row and a primary key index.

Migration provided was hand-written and tested to be able to move
in both directions.