Add `yaml_quote` function
It is equivalent to `shlex.quote` but for YAML. Takes any
arbitrary Python object that can be converted to YAML such
as dictionary, list, integer, float or string  and returns
a valid YAML single line string. The strings would be converted
to double quoted scalars.

Will be used for template filters.

This is implemented using the Dumper class of the `yaml` module
and the StringIO class from `io`. The Dumper is configured to use
flow style double quoted YAML objects in to a buffer. The buffer
is parsed in to string and then a leading YAML object indicator
`--- ` and trailing newline is cut to create a string that can be
embedded in the templates without any newlines. The buffer is
reused between every quote which improves the performance.

When C YAML extension is available it will be used.

Also add unit tests to verify behavior including quoting
dictionaries, unprintable characters and extremely long lines.
Test the pure Python implementation of `yaml` if C implementation
has been tested.