Add `yaml_quote` and `shlex_quote` filters to Device and DeviceType template environments
**This is a backwards compatible change** because the new filters
are not used in any templates.

`yaml_quote` takes any Python object that can be converted to
safe YAML and returns a string that can be embedded in YAML.

`shlex_quote` takes a string and returns a string that will be
interpreted by a POSIX shell as a single string.

Both filters can be used as filter sections (`{% filter %}`) and
filter pipeline. (`{{ object|yaml_quote }}`)