lava_dispatcher: fix function retry decorator avh api calls
For avh API calls like `v1_delete_image`, the retry decorator should
retry the call on `AvhApi.ApiException` and return on `NotFoundException`
which is a subclass of `ApiException`.

Because the expected exception is captured first, when its scope is too
big, the retry may never get triggered. This change sets the expected
exception to `None` and only tries to capture it when it is defined.

Exception is raised if the exception for function retry is a subclass
of expected exception. In this case, the function will never be retried.

AVH api return type check is already covered by the avh-api library
itself. There is no need to do the check again in LAVA.

Signed-off-by: Chase Qi <chase.qi@linaro.org>