qemu.jinja2: Fix indentation, vga and serial options and kvm enabled everywhere
Fix lists indentations. Particularly the `available_architectures`
and `options`.

Fix `vga` and `serial` options only being available when emulating
ARM architecture. Also fix those options not prepending `-vga` and
`-serial` to passed values.

Only enable KVM when the host supports it. For x86_64 which is
assumed in `qemu.jinja2` only architectures that support it are
`i386`, `amd64` and `x86_64`. For Aarch64 in `qemu-aarch64.jinja2`
those are `aarch64`, `arm64` and `arm`.