lava_dispatcher: save raw pexpect output with universal newline
pexpect replaces `\n` with `\r\n`. Newlines in pexpect output end up in
two forms, `\r\n` and `\r\r\n`. The former is converted from `\n` while
the latter is converted from `\r\n`.

In lava dispatcher, we can simply split pexpect output by either
`\r\r\n` or `\r\n`. In this way, lava saves pexpect output as it is but
with unified Unix style newline character `\n`.

Single `\r` character is saved as it is in case it is needed outside of
lava.

Signed-off-by: Chase Qi <chase.qi@linaro.org>