lava_dispatcher: use pipeline initial max retries if it is bigger
Device level default max retires set in `base.jinja2` is 1. It overwrites
the pipeline initial value whenever it is set.

To make it possible to bump the initial pipeline level max retires for
network related action like `DownloaderAction`, this change compares the
retry numbers and uses the bigger one for the final run.

Signed-off-by: Chase Qi <chase.qi@linaro.org>