lava_dispatcher: Import boot strategy module based on method name
Instead of having a strategy module that imports everything in to
one module and then gets imported by parser keep a dictionary of
boot method name to a module name. During the boot method selection
import the required module.

This makes certain dependencies no longer be required even if
the jobs are not using them. (for example AVH) Also memory use
should get reduced as well as faster startup.