tests: Mock compression downloads instead of using Linaro servers
Right now the downloads decompression tests have to download the
compressed data from `images.validation.linaro.org` server every
time the test is ran.

This creates several problems:

* What if images.validation.linaro.org goes down? No CI pipelines
  will be able to complete.
* Because it is using unencrypted HTTP protocol it is probably
  possible to create man-in-the-middle attacks against the used
  decompression subprocesses.
* It is also slow and takes a few second to complete.

Instead use the `responses` library that can mock the requests
library. It is already packaged for Debian under `python3-responses`
name so it is easy to add. Switch domain name used to `example.com`
to have the tests fail fast in case when mock fails to patch.

The downloaded files from `images.validation.linaro.org` are
located in the `tests/lava_dispatcher/download_artifacts/` directory.
Originally the decompression tests were added in 51c81f6598793c7c7002a0287c8d1f04497ac791
so the `.license` files indicate the same author as of that commit.

Also reorganize the TestDecompression class.

The `test_multiple_decompressions` doesn't actually download
anything and is misleadingly named. Rename it to
`test_download_decompression_map` to indicate what
it actually tests and move it to separate class.

Move `test_bad_download_decompression` in to the TestDecompression
class because it also relies on the requests mocking so by moving
it in to same class will allow to reuse the mock setup code.

The sample job are now created inside the test instead of the
setUp() methods because `test_download_decompression` and
`test_bad_download_decompression` use different job definitions.
Also the `job.validate()` makes HTTP calls to it has to be
run after the mock is setup.