Use quote filters for base-uboot.jijna2's setenv bootargs
Currently that template only puts the boot args in to single
quotes and the setenv command in to double quotes.
However, this is not enough. For example, if your job's
`extra_kernel_args` contains double quotes the produced YAML will
be invalid and the job will fail.

Instead of manual quoting use the new filter functions `shlex_quote`
and `yaml_quotes`. This way the produced YAML is always valid.