lava_dispatcher: fix shell logger for pexpect input
We assumed that pexpect always use `\r\n` for newlines. That is only
true for output. Input is a different case, `os.linesep()` that varies
by platform is used for sending new lines. It is `\n` on Unix-like
systems.

LAVA `sendline` method uses `\r\n` for UEFI that type set to `dos` and
uses `\n` for everything else.

shell logger prints lines before the last `\n` and keeps the remaining
line in memory for the next iteration. Lines are split by `\r\r\n` or
`\r\n` now. This works fine for DUT output, but input lines sent by
lava end up with `\n` will be eaten...

This change adds `\n` back as a new line character for splitting log
lines. It also uses `\n` for the `flush` method to unify LAVA input.

Signed-off-by: Chase Qi <chase.qi@linaro.org>