rest-0.3: Initial implementation of TestJob REST API v0.3
Initial serializer and read-only viewset.

The serializer implements the `error_msg` and `error_type`
fields from the XMLRPC's `list` method. All fields are read-only
because in the future the job creation will be implemented using
a details route instead of restframework's magic.

This implements a different organization of REST API files.
Instead of having all serializers in single file or all viewsets
in a single file, put everything related to TestJob (serializers,
viewsets, filters) in to a `testjob.py` file. This makes it easier
to work with as related classes will be in a single file.