tests: Fix test_api.py tests failing because of ordering
The issue is that the tests assumed the database would always return
entities in the same order they were created which is sometimes
not a case.

This would cause the test to fail with the following message:

AssertionError: assert 'qemu' == 'b2260'

The Device and DeviceType objects do not have and id field which
can be used to sort them by creation sequence so instead use their
primary keys "hostname" and "name" respectively.