Remove Debian 10 django.urls.re_path workaround
The `django.urls.re_path` was added in Django 2.0 but Debian 10
shipped the Django 1.11. In Debian 10 the equivalent to the
`re_path` has to imported from the `django.conf.urls`.

Since the lowest supported Debian version is now Debian 11 the
workaround can be removed and all urls can be configured directly
with `django.urls.re_path`.