tests: Add ShellLogger tests
It tests how ShellLogger will handle the data mangling that pseudo
TTY does. The pexpect runs its commands in pseudo TTY which alters
the output. For example, each `\n` will get expanded to `\r\n`.
This means if the command outputs the `\r\n` the pseudo TTY will
change it to `\r\r\n` but the ShellLogger must log it as a single
line.

Also the input gets logged twice. First time by pexpect which is
not mangled and second time by pseudo TTY which get mangled.
Test this behavior as well.